Move app-data-hash into app-data crate
