--- crates/app-data-hash/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "app-data-hash"
-version = "0.1.0"
-authors = ["Cow Protocol Developers <dev@cow.fi>"]
-edition = "2021"
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-hex-literal = { workspace = true }
-tiny-keccak = { version = "2.0.2", features = ["keccak"] }
-
-[lints]
-workspace = true
