+++ crates/model/Cargo.toml
@@ -11,7 +11,6 @@ doctest = false
 [dependencies]
 anyhow = { workspace = true }
 app-data = { path = "../app-data" }
-app-data-hash = { path = "../app-data-hash" }
 bytes-hex = { path = "../bytes-hex" }
 bigdecimal = { workspace = true }
 chrono = { workspace = true, features = ["serde", "clock"] }
